TwitterMining/requirements.txt
Flask==1.0.2
Werkzeug<0.13.0,>=0.12.0
tweepy==3.6.0
google-cloud-datastore==1.7.0
google-api-python-client
requests-toolbelt==0.8.0
gunicorn==19.7.1